CI: Inline Core/DX approval check in Dangerfile
Remove the separate CommonJS helper and unit tests; keep the approval
gate logic directly in scripts/dangerfile.js.
